requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
autopep8==2.0.2
awsebcli==3.20.6
beautifulsoup4==4.12.2
botocore==1.29.99
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
crispy-bootstrap3==2022.1
cryptography==40.0.2
cssselect2==0.7.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.2
django-admin-honeypot==1.1.0
django-admin-notification==0.2
django-adminlte-3==0.1.6
django-allauth==0.54.0
django-appconf==1.0.5
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==23.1
django-ckeditor==6.5.1
django-classy-tags==4.0.0
django-cms==3.11.3
django-cookie-consent==0.3.1
django-crispy-forms==2.0
django-decouple==2.1
django-formtools==2.4
django-icons==23.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-leaflet==0.29.0
django-mptt==0.14.0
django-paypal==2.0
django-recaptcha==3.0.0
django-rest-framework==0.1.0
django-rest-paystack==2.2.2
django-sekizai==4.0.0
django-shortuuidfield==0.1.3
django-taggit==3.1.0
django-tinymce==3.6.1
django-treebeard==4.7
django-widget-tweaks==1.4.12
django-xhtml2pdf==0.0.4
djangocms-admin-style==3.2.4
djangocms-owl==0.1.11
djangorestframework==3.14.0
filelock==3.11.0
fontawesomefree==6.4.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
jmespath==1.0.1
jsonfield==3.1.0
lxml==4.9.2
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.1
pathspec==0.10.1
Pillow==9.5.0
platformdirs==3.2.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyHanko==0.18.0
pyhanko-certvalidator==0.22.0
PyJWT==2.6.0
pypdf==3.8.1
pypiwin32==223
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==5.4.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.26.0
requests-oauthlib==1.3.1
semantic-version==2.8.5
shortuuid==1.0.11
six==1.14.0
soupsieve==2.4.1
sqlparse==0.4.3
svglib==1.5.1
termcolor==1.1.0
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
Unipath==1.1
uritools==4.0.1
urllib3==1.26.15
virtualenv==20.21.0
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==6.4.0
xhtml2pdf==0.2.10
